Fix for #3665

+++ wled00/cfg.cpp
@@ -696,7 +696,7 @@ void serializeConfig() {
 
   JsonObject wifi = root.createNestedObject("wifi");
   wifi[F("sleep")] = !noWifiSleep;
-  wifi[F("phy")] = (int)force802_3g;
+  wifi[F("phy")] = force802_3g;
 
   #ifdef WLED_USE_ETHERNET
   JsonObject ethernet = root.createNestedObject("eth");
